add the difference of the edit cards components when they output on the lead edit page
