+++ client/src/components/editLead/editLeadSidebar/EditCard/EditCard.js
@@ -8,44 +8,19 @@ import MainField from "./CardFields/MainField";
 
 class EditCard extends Component {
 
-    constructor(){
-        super();
-        this.state = {
-            leadId: '',
-            leads: {}
-        }
-    }
-
-    componentWillReciveProps(nextProps){
-        if(nextProps){
-            this.setState({
-                leadId: nextProps,
-                leads: {}
-            })
-        }
-    }
-
-    componentDidMount(){
-        let id = '5b6d7b35401106620b1a1a6d';
-        this.props.loadLead(id);
-    }
-
     render() {
-        const { leads } = this.props.leads;
-        if(!Object.values(leads).length) {
-            return <div />;
-        }
-        const fields = leads.contact.custom.map((field) =>
+        let title = ('organization' in this.props.value) ? 'Person' : 'Organization';
+        const fields = this.props.value.custom.map((field) =>
             <CardField fieldValues={Object.values(field)} fieldName={Object.keys(field)}/>);
         return (
             <div className={styles.container}>
                 <div className={styles.title}>
                     <span className={styles.titleName}>
-                        Person
+                        {title}
                     </span>
                 </div>
                 <div className={styles.fields}>
-                    <MainField contact={leads.contact}/>
+                    <MainField value={this.props.value}/>
                     {fields}
                 </div>
             </div>
@@ -53,13 +28,4 @@ class EditCard extends Component {
     }
 }
 
-EditCard.propTypes = {
-    loadLead: PropTypes.func.isRequired,
-    leads: PropTypes.object.isRequired
-};
-
-const mapStateToProps = (state) => ({
-    leads: state.leads
-});
-
-export default connect(mapStateToProps, { loadLead })(EditCard);
+export default EditCard;
